+++ config/routes.rb
@@ -41,6 +41,8 @@ Rails.application.routes.draw do
       resources :adverse_event_comments
     end
 
+    resources :categories
+
     resources :contacts
     resources :documents do
       member do